+++ controller/src/main/java/itacademy/AddTeacher.java
@@ -4,6 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -11,6 +12,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+@WebServlet(value = {"/admin/addteacher"})
 public class AddTeacher extends HttpServlet {
     private final static Logger log = LoggerFactory.getLogger(AddTeacher.class);
