+++ controller/src/main/java/itacademy/filters/AuthFilter.java
@@ -1,11 +1,13 @@
-package itacademy;
+package itacademy.filters;
 
 import javax.servlet.*;
+import javax.servlet.annotation.WebFilter;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+@WebFilter(value = {"/admin/*"})
 public class AuthFilter implements Filter {
 
     @Override
