+++ controller/src/main/java/itacademy/filters/EncodingFilter.java
@@ -1,8 +1,10 @@
-package itacademy;
+package itacademy.filters;
 
 import javax.servlet.*;
+import javax.servlet.annotation.WebFilter;
 import java.io.IOException;
 
+@WebFilter(value = {"/*"})
 public class EncodingFilter implements Filter {
     @Override
     public void init(FilterConfig filterConfig) throws ServletException {
