+++ view/src/main/java/itacademy/Salaries.java
@@ -1,6 +1,7 @@
 package itacademy;
 
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -8,9 +9,9 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.List;
 
+@WebServlet(value = {"/admin/salaries"})
 public class Salaries extends HttpServlet {
     private static List<Teacher> teachers = TeachersRepository.getTeachers();
-
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
