+++ view/src/main/java/itacademy/Teachers.java
@@ -1,6 +1,7 @@
 package itacademy;
 
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -8,6 +9,7 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.List;
 
+@WebServlet(value = {"/admin/teachers"})
 public class Teachers extends HttpServlet {
     private static List<Teacher> teachers = TeachersRepository.getTeachers();
