+++ view/src/main/webapp/WEB-INF/web.xml
@@ -15,67 +15,12 @@
     </servlet-mapping>
 
     <servlet>
-        <servlet-name>addteacher</servlet-name>
-        <servlet-class>itacademy.AddTeacher</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>addteacher</servlet-name>
-        <url-pattern>/admin/addteacher</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>teachers</servlet-name>
-        <servlet-class>itacademy.Teachers</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>teachers</servlet-name>
-        <url-pattern>/admin/teachers</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>salaries</servlet-name>
-        <servlet-class>itacademy.Salaries</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>salaries</servlet-name>
-        <url-pattern>/admin/salaries</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
         <servlet-name>auth</servlet-name>
         <jsp-file>/auth.jsp</jsp-file>
     </servlet>
     <servlet-mapping>
         <servlet-name>auth</servlet-name>
         <url-pattern>/auth</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>index</servlet-name>
-        <servlet-class>itacademy.Index</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>index</servlet-name>
-        <url-pattern>/</url-pattern>
-        <url-pattern>/index*</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>login</servlet-name>
-        <servlet-class>itacademy.Login</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>login</servlet-name>
-        <url-pattern>/login</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>logout</servlet-name>
-        <servlet-class>itacademy.Logout</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>logout</servlet-name>
-        <url-pattern>/logout</url-pattern>
     </servlet-mapping>
 
     <servlet>
@@ -86,26 +31,6 @@
         <servlet-name>admin</servlet-name>
         <url-pattern>/admin</url-pattern>
     </servlet-mapping>
-
-
-
-    <filter>
-        <filter-name>EncodingFilter</filter-name>
-        <filter-class>itacademy.EncodingFilter</filter-class>
-    </filter>
-    <filter-mapping>
-        <filter-name>EncodingFilter</filter-name>
-        <url-pattern>/*</url-pattern>
-    </filter-mapping>
-
-    <filter>
-        <filter-name>AuthFilter</filter-name>
-        <filter-class>itacademy.AuthFilter</filter-class>
-    </filter>
-    <filter-mapping>
-        <filter-name>AuthFilter</filter-name>
-        <url-pattern>/admin/*</url-pattern>
-    </filter-mapping>
 
 
     <jsp-config>
